+++ go/client/chat_cli_utils.go
@@ -133,3 +133,25 @@ func (f chatCLIInboxFetcher) fetch(ctx context.Context, g *libkb.GlobalContext) 
 
 	return res.Conversations, nil
 }
+
+func fetchOneMessage(g *libkb.GlobalContext, conversationID chat1.ConversationID, messageID chat1.MessageID) (chat1.MessageFromServerOrError, error) {
+	deflt := chat1.MessageFromServerOrError{}
+
+	chatClient, err := GetChatLocalClient(g)
+	if err != nil {
+		return deflt, err
+	}
+
+	arg := chat1.GetMessagesLocalArg{
+		ConversationID: conversationID,
+		MessageIDs:     []chat1.MessageID{messageID},
+	}
+	res, err := chatClient.GetMessagesLocal(context.TODO(), arg)
+	if err != nil {
+		return deflt, err
+	}
+	if len(res.Messages) < 0 {
+		return deflt, fmt.Errorf("empty messages list")
+	}
+	return res.Messages[0], nil
+}
